Don't freeze values set via setUnfrozen.
Freezing values may not always be desired. For example, Doctrine closes its
entity manager if an error occurs on flush. The recommended way to recover
from this is to replace the entity manager with a new one, while reusing the
configuration. Pimple's freezing feature gets in the way.

This change preserves backwards compatibility. It's up to authors to decide
whether values are frozen, while defaulting to freezing.

It's debatable whether it should be allowed to disable freezing.  According
to PSR-11 however, a "user SHOULD NOT rely on getting the same value on 2
successive calls".  This change is a compromise, it offers freezing by default,
but also allows disabling it per service.